move segmentReader(IndexReader) util to Lucene.java
